(feat) Begin virtual interface management
The following begins the work needed to flush out the virtual interfaces
that OpenBSD supports.

The first is, CARP.  We beign with a a wrapper class
bsd::network::interface::carp that has the needed paramaters.  We chew
on the inputs, before we spit out the ifconfig lines that we can pass to
the bsd::network::interface define to create the carp interface.  Works
like a charm.

Next up is the GRE interface managemnt.  We just piggy back on the
bsd::network::interface type here, and enable GRE when its needed.  I
expec that this interface will change in the future.  Ideally, pass in
the needed values similar to how the carp interface is done.  For now
this works, and is worth commiting.  Example has been inclued in the
tests/ directory.

Sysctl classes for enablement of each of the above has been included.

Add some basic testing for the above as well.