Fix vlan(4) interface address asignment
Previously all of the settings were being set all on one line.  This
cased issues when the address assignment came before the VLAN
assignment.

This work breaks the VLAN assignment and the address assignment out into
multiple lines, the VLAN assignment being the first of which.

I expect that this pattern should also be applied to the carp and other
virtual interfaces that require address assignment only after the
interface options are set.