The vether remains unprocessed
Adding vether to the supported list means that the interface is
expected to have already been processed.  This is not quite so.  Until
we build a bsd::network::interface::vether define, we need not include
vether in the supported list, as implementing an interface with vether
does not require it.  The bsd::network::interface can be used directly.