Add type and provider for create and destroy
This work implements a simple type and provider to handle the ifconfigs
necessary to manage the state of an interface.  For physical
(non-pseudo) devices, this means the ability to bring the interface up
with 'ifconfig up' and down with 'ifconfig down'.  For pseudo devices
(called clonable on FreeBSD), this means calling 'ifconfig create' and
'ifconfig destroy' as required.

This should allow the manifests to make the correct call depending on
the ensure parameter to pass to a bsd_interface resource for state
management.