Improve logic for exists? method
Test the logic of what the exists method should do when querying the
state of real and pseudo interfaces.  Pseudo interfaces can be
destroyed, and so their mere existence should mean that they are
present.  Pseudo interfaces are only detected absent when they are not
found on the system.  The states up, down, and present for pseudo
interfaces all yield 'present' state.  This logic differs on real
interfaces, since real interfaces cannot be destroyed, their down state
also counts as absent, where only present and up yield the 'present'
state.