Notify the netif resource to restart on change
Without this change, the interface will only match the state set by
ensure, but the FreeBSD netif system is not notified and thus no
interface addresses are actually set.

This change ensures that when the rc.conf line changes for a given
interface, that the netif script is executed with the 'restart' option
for the given interface, and the addresses requested are applied to the
interface.