Add a parents parameter to bsd::network::interface. This can be a string or an array. The value of that parameter is passed to Bsd_interface.rb when the interface is started.
Add a parents parameter to bsd_interface.rb, and autorequire
the parents. A bit of validation etc. of the interface name
done.

In bsd::network::interface, require the creation of hostname.if
for OpenBSD or the shell_config for FreeBSD to be done before
the Bsd_interface is created.

The subclasses in bsd::network::interface::* where parent interfaces
make sense to set, i.e. bridge, carp, trunk and vlan, deduct the
parent interface based on the interface or device parameter, and
hand that over to the parents parameter of bsd::network::interface.

Specs added for bsd::network::interface::* defined types, and
the parents defined types.

Further added some flesh to the README.md mentioning the interface
ordering and the parents parameter of the bsd::network::interface
defined type.