Add special defined type for pfsync interfaces bsd::network::interface::pfsync
This new type makes use of the parent interface of
bsd::network::interface, setting the value of the syncdev
parameter to the parent.

README.md updated, and spec tests added for the new defined
type and the corresponding hostname_if function.

This defined type is a bit special compared to the others,
since, when no parameters given, it explicitly sets the defaults

i.e. just create a pfsync0 interface, without parameters will end
up with a /etc/hostname.pfsync0 with:

-syncdev -syncpeer maxupd 128 -defer
up

What is default anyways is taken from the pfsync manual page.

This has the advantage that interfaces are easier to reconfigure.
Instead of ensuring its absense, and re-creating it, a parameter
can just be removed, and it will be reset to the default value.