Fixup carp with multiple addresses, and add specs to test for that as well as adding specs for giving multiple values to carp interfaces.
When there were multiple addresses given as an array, they were
only added onto a single line, breaking sh /etc/netstart carpXXX

Now correctly joining the data in lib/puppet_x/bsd/hostname_if/carp.rb
with a "\n" instead of the wrong ' '.
While there, I reordered the lines, like to have the same order
as done in the trunk interface.

Updating spec tests, to deal with that new order of lines, as well
as adding a test to check for IP addresses given as array as well
as values parameter given as array.