Update module dependencies
This module now has three extra module dependencies, and drops
shell_config as a requirement.  The augeas providers are under active
development, and the providers are growing for a number of other types.
As such, to avoid needing two different modules that fundamentally do
the same thing, this change is my preference.

To update, simply have a look at the new dependencies and install the
needed modules.  If you are currently using shell_config elsewhere in
your environment, there will be no conflict, so you may go about your
day.  If you are not using shell_config elsewhere in your environment,
updating to this module would prove a reasonable time to remove stop
including it on your module list.